pyproject.toml
[project]
name = "mosaicml-streaming-torchrun"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mosaicml-streaming>=0.12.0",
    "torch>=2.7.0",
    "torchvision>=0.22.0",
    "tqdm>=4.67.1",
]
